% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usescape_PACKAGE.R
\name{clust_stack}
\alias{clust_stack}
\title{Back-association of population-level clustering to individual clusters}
\usage{
clust_stack(ls1, pop, ind, table, min_fix = 3)
}
\arguments{
\item{table}{The output of table_cluster}

\item{grid}{The output of the loop function}

\item{pop_clust}{The output of the pop_clust function}

\item{ind_clust}{The output of the ind_clust function}
}
\value{
A list of raster stack object.
}
\description{
Generate individual-level rasters of population-level clustering. For each individual, the function generates a raster stack containing a raster of the most likely cluster, and several rasters giving the probability of observing each cluster.
}
\examples{
data(puechcirc)
traj<-na.omit(puechcirc)
ls1<-loop_id(traj, res=300)
table<-table_cluster(traj, ls1)
ind<-ind_clust(table)
pop<-pop_clust(traj, ind)
stack<-clust_stack(ls1, pop, ind, table, min_fix = 3)
plot(stack[[1]][[1]]) #Plot first individuals 
plot(stack[[2]][[1]]) #Plot second individuals 
}
\keyword{ind_clust}
\keyword{loop_id}
\keyword{pop_clust}
\keyword{table_cluster}
\keyword{timing2stack}
\keyword{traj2timing}
