% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usescape_PACKAGE.R
\name{ind_clust}
\alias{ind_clust}
\title{Individual-level clustering of intensity of use metrics}
\usage{
ind_clust(
  table,
  nb_clust = 1:5,
  min_fix = 3,
  vars = c("Number.visits", "Total.duration", "Mean.duration", "CV.duration",
    "Mean.interval", "SD.interval")
)
}
\arguments{
\item{table}{An output from the table_cluster function}

\item{nb_clust}{The number of clusters to be tested, see the documentation for mclust for more information. Default = 1:5.}

\item{min_fix}{The minimum of locations in a pixel needed to be included in the analysis. Default=3.}

\item{vars}{The variable to be included. Default = c("Number.visits", "Total.duration", "Mean.duration", "CV.duration", "Mean.interval", "SD.interval")}
}
\value{
A list object with each element representing an individual.
}
\description{
Perform individual-level clustering (first step) of intensity of use metrics. This function uses the output of table_cluster and perform a mixture-model. Users can select which variables will be used and the maximum number of clusters. See also mclust
}
\examples{
data(puechcirc)
traj<-na.omit(puechcirc)
ls1<-loop_id(traj, res=300)
table<-table_cluster(traj, ls1)
ind<-ind_clust(table)
pop<-pop_clust(traj, ind)
stack<-clust_stack(ls1, pop, ind, table, min_fix = 3)
plot(stack[[1]][[1]]) #Plot first individuals 
plot(stack[[2]][[1]]) #Plot second individuals 
}
\keyword{loop_id}
\keyword{table_cluster}
\keyword{timing2stack}
\keyword{traj2timing}
