% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usescape_PACKAGE.R
\name{loop_id}
\alias{loop_id}
\title{Looping over all individuals}
\usage{
loop_id(traj, res = 100)
}
\arguments{
\item{traj}{An object produce by the function adehabitatLT with multiple individuals}

\item{res}{Grid size, will be apply to all individuals}
}
\value{
A list object containing a raster stack object for each individual
}
\description{
Extract the timing history and intensity of use metrics for all individuals in a traj object
}
\examples{
data(puechcirc)
traj<-na.omit(puechcirc)
ls1<-loop_id(traj, res=300)
table<-table_cluster(traj, ls1)
ind<-ind_clust(table)
pop<-pop_clust(traj, ind)
stack<-clust_stack(ls1, pop, ind, table, min_fix = 3)
plot(stack[[1]][[1]]) #Plot first individuals 
plot(stack[[2]][[1]]) #Plot second individuals 
}
\keyword{timing2stack}
\keyword{traj2timing}
