% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usescape_PACKAGE.R
\name{mask_uncertain}
\alias{mask_uncertain}
\title{Mask uncertain pixel from classification}
\usage{
mask_uncertain(out, p = 0.05)
}
\arguments{
\item{out}{output of clust_use}

\item{p}{uncertainty threshold (values with uncertainty higher than p will be masked). Default = 0.05}
}
\value{
return a raster object
}
\description{
Remove pixel with uncertainty higher than a specific value from a plot
}
\details{
Remove pixel with uncertainty higher than a specific value from a plot
}
\examples{
traj1<-sim_mov(type="OU", npatches=3, grph=T)
timing_ls<-traj2timing(traj1, res=50, grid=NULL)
stck<-timing2stack(timing_ls, col=1,2,3,4,6,7) 
test<-clust_use(stck)
plot(test[[2]], col=c("grey", "blue", "green", "red", "purple", "orange"))
plot(test[[3]]) #Uncertainty 
mask<-mask_uncertain(test, p=0.01)
plot(mask, col=c("grey", "blue", "green", "red", "purple", "orange"))

}
