% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usescape_PACKAGE.R
\name{pop_clust}
\alias{pop_clust}
\title{Population-level clustering of intensity of use metrics}
\usage{
pop_clust(traj, ls, n_clust = 1:5)
}
\arguments{
\item{traj}{The trajectory object}

\item{ls}{Individual-level clustering object, the output of ind_clust.}

\item{n_clust}{The number of clusters to be tested, see the documentation for mclust for more information. Default = 1:5.}
}
\value{
A list object with each element representing an individual.
}
\description{
Combine individual-level clustering of movement metrics into a population-level clustering (second step). Users can  define the number of clusters. See also mclust
}
\examples{
data(puechcirc)
traj<-na.omit(puechcirc)
ls1<-loop_id(traj, res=300)
table<-table_cluster(traj, ls1)
ind<-ind_clust(table)
pop<-pop_clust(traj, ind)
stack<-clust_stack(ls1, pop, ind, table, min_fix = 3)
plot(stack[[1]][[1]]) #Plot first individuals 
plot(stack[[2]][[1]]) #Plot second individuals 
}
\keyword{atiming2stack}
\keyword{ind_clust}
\keyword{loop_id}
\keyword{table_cluster}
\keyword{traj2timing}
