% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usescape_PACKAGE.R
\name{res_test}
\alias{res_test}
\title{Test several pixel sizes (resolution) and estimate the coefficient of variation in residency time}
\usage{
res_test(mov, res_seq = c(50, 100, 150), unit_time = "secs")
}
\arguments{
\item{mov}{simulated movement, replace with trajectory object of interest}

\item{res_seq}{a vector of pixel size.}

\item{unit_time}{time unit for parameters}
}
\value{
a graph showing the coefficient of variation in residency time as a function of pixel size.
}
\description{
Test several pixel sizes (resolution) and estimate the coefficient of variation in residency time
}
\details{
The function estimates the coefficient of variation in residency time calculated over grid of varying pixel size and produce a graph showing their values
}
\examples{
code examples using function - these are vital ###########################
traj1<-sim_mov(type="OU", npatches=3, grph=T)
res_test(traj1, res_seq=c(50,100,150))
}
