% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usescape_PACKAGE.R
\name{table_cluster}
\alias{table_cluster}
\title{Convert a list of timing2stack object to a data.frame for clustering}
\usage{
table_cluster(traj, ls)
}
\arguments{
\item{traj}{The trajectory used in loop (a traj object)}

\item{ls}{The output of the loop_id function}
}
\value{
A data.frame object.
}
\description{
Convert output of loop_id function to a data.frame.
}
\examples{
data(puechcirc)
traj<-na.omit(puechcirc)
ls1<-loop_id(traj, res=300)
table<-table_cluster(traj, ls1)
ind<-ind_clust(table)
pop<-pop_clust(traj, ind)
stack<-clust_stack(ls1, pop, ind, table, min_fix = 3)
plot(stack[[1]][[1]]) #Plot first individuals 
plot(stack[[2]][[1]]) #Plot second individuals  
}
\keyword{loop_id}
\keyword{timing2stack}
\keyword{traj2timing}
