% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usescape_PACKAGE.R
\name{timing2stack}
\alias{timing2stack}
\title{Timing history to raster stack conversion}
\usage{
timing2stack(timing_ls, unit_time = "secs")
}
\arguments{
\item{timing_ls}{timing history for each pixel}

\item{unit_time}{time unit for timing_ls parameter}
}
\value{
a raster stack object
}
\description{
Extract use metrics from timing history
}
\details{
This function extract the intensity of use metrics from the timing history and return a stack object of a raster layer for each metric.
}
\examples{
 
traj1<-sim_mov(type="OU", npatches=3, grph=T)
timing_ls<-traj2timing(traj1, res=50, grid=NULL)
stck<-timing2stack(timing_ls) 
plot(stck)

}
