% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usescape_PACKAGE.R
\name{traj2timing}
\alias{traj2timing}
\title{Calculation of timing history from movement data}
\usage{
traj2timing(mov, res = 100, grid = NULL)
}
\arguments{
\item{mov}{Movement trajectory, need to be a ltraj object}

\item{res}{Grid size (based on coordinate system of movement trajectory)}

\item{grid}{User specified grid (a raster), needs to have a larger extent than the movement trajectory}
}
\value{
A list of objects containing the timing history and the grid use.
}
\description{
Transform an ltraj object to a timing history (entry and exit time in each pixel) using a user-specified grid size.
}
\details{
Using a specific grid cell size, this function return the timing history (entry and exit time in each pixels) based on an individual trajectory.
}
\examples{
traj1<-sim_mov(type="OU", npatches=3, grph=T)
timing_ls<-traj2timing(traj1, res=50, grid=NULL)

}
\keyword{timing}
